refactor: extract utilities to dedicated packages
- Move csp-analyzer to packages/csp
- Move login-detector, privacy-finder, tos-finder to packages/detectors
- Add browser-adapter utility
- Update extension imports to use new packages
