+++ app/extension/.wxt/types/imports.d.ts
@@ -1,12 +1,13 @@
 // Generated by wxt
 export {}
 declare global {
-  const CSPAnalyzer: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-analyzer').CSPAnalyzer
   const CSPReporter: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-reporter').CSPReporter
   const ContentScriptContext: typeof import('wxt/utils/content-script-context').ContentScriptContext
   const InvalidMatchPattern: typeof import('wxt/utils/match-patterns').InvalidMatchPattern
   const MatchPattern: typeof import('wxt/utils/match-patterns').MatchPattern
   const browser: typeof import('wxt/browser').browser
+  const browserAdapter: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/browser-adapter').browserAdapter
+  const createBrowserAdapter: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/browser-adapter').createBrowserAdapter
   const createIframeUi: typeof import('wxt/utils/content-script-ui/iframe').createIframeUi
   const createIntegratedUi: typeof import('wxt/utils/content-script-ui/integrated').createIntegratedUi
   const createShadowRootUi: typeof import('wxt/utils/content-script-ui/shadow-root').createShadowRootUi
@@ -15,11 +16,8 @@ declare global {
   const defineContentScript: typeof import('wxt/utils/define-content-script').defineContentScript
   const defineUnlistedScript: typeof import('wxt/utils/define-unlisted-script').defineUnlistedScript
   const defineWxtPlugin: typeof import('wxt/utils/define-wxt-plugin').defineWxtPlugin
-  const detectLoginPage: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/login-detector').detectLoginPage
   const fakeBrowser: typeof import('wxt/testing').fakeBrowser
-  const findPrivacyPolicy: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/privacy-finder').findPrivacyPolicy
   const injectScript: typeof import('wxt/utils/inject-script').injectScript
-  const isLoginPage: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/login-detector').isLoginPage
   const onCookieChange: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/cookie-monitor').onCookieChange
   const startCookieMonitor: typeof import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/cookie-monitor').startCookieMonitor
   const storage: typeof import('wxt/utils/storage').storage
@@ -58,15 +56,6 @@ declare global {
   export type { CookieChangeCallback } from '/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/cookie-monitor'
   import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/cookie-monitor')
   // @ts-ignore
-  export type { CSPAnalyzer, DomainCSPPolicy, GeneratedCSPByDomain } from '/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-analyzer'
-  import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-analyzer')
-  // @ts-ignore
   export type { CSPReporter, ReportPayload } from '/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-reporter'
   import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/csp-reporter')
-  // @ts-ignore
-  export type { LoginDetectionResult } from '/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/login-detector'
-  import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/login-detector')
-  // @ts-ignore
-  export type { PrivacyPolicyResult } from '/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/privacy-finder'
-  import('/Users/hikae/ghq/github.com/HikaruEgashira/service-policy-auditor/app/extension/utils/privacy-finder')
 }
