+++ app/extension/entrypoints/background.ts
@@ -17,8 +17,8 @@ import type {
   CSPGenerationOptions,
 } from "@service-policy-auditor/core";
 import { DEFAULT_CSP_CONFIG } from "@service-policy-auditor/core";
+import { CSPAnalyzer, type GeneratedCSPByDomain } from "@service-policy-auditor/csp";
 import { startCookieMonitor, onCookieChange } from "@/utils/cookie-monitor";
-import { CSPAnalyzer, type GeneratedCSPByDomain } from "@/utils/csp-analyzer";
 import { CSPReporter } from "@/utils/csp-reporter";
 
 const MAX_EVENTS = 1000;
