+++ packages/csp/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@service-policy-auditor/csp",
+  "version": "0.0.1",
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "dependencies": {
+    "@service-policy-auditor/core": "workspace:*"
+  }
+}
